devops/prometheus/service_discovery/requirements.txt
PyYAML==6.0
Jinja2==3.1.2
pymongo==3.12.0
requests==2.25.1
Flask==2.1.3